api/requirements.txt
llama-index-llms-ollama
fastapi[standard]
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv